refactor: refactor webhooks controller
