--- lib/media_server_web/controllers/webhook_movie_controller.ex
@@ -1,15 +1,0 @@
-defmodule MediaServerWeb.Webhooks.MovieController do
-  use MediaServerWeb, :controller
-
-  def create(conn, %{"eventType" => "Download"}) do
-    MediaServer.MoviesIndex.reset()
-
-    conn
-    |> send_resp(201, "Ok")
-  end
-
-  def create(conn, _payload) do
-    conn
-    |> send_resp(200, "Ok")
-  end
-end
